# swagger.yaml
openapi: 3.0.3
info:
  title: SET Service API
  version: 1.1.0
  description: |
    SET Service REST API for Worker, Company, Super Admin, and restricted Admin roles.

    Core worker flow:
    Worker registers -> verifies OTP -> becomes pending_approval -> super admin or permitted admin approves or rejects -> approved worker can log in and receive tokens.

    Contract-wide observability and authentication rules:
    - Every HTTP response includes `X-Request-ID` and `X-Correlation-ID`.
    - Clients may send either header using 8-128 characters from `[A-Za-z0-9._:-]`; invalid values are replaced server-side.
    - Error bodies include the effective `request_id`; the correlation identifier remains in the response header.
    - `BearerAuth` accepts access tokens only. Refresh tokens are valid only in the documented refresh/logout request bodies.
    - OTP request success means the request was accepted for delivery; it does not prove arrival at the handset or mailbox.
    - Interactive Swagger UI is disabled in production and should be exposed only in controlled non-production environments.
servers:
  - url: http://localhost:3000/v1
    description: Local development
  - url: https://api.setservice.az/v1
    description: Production
security: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT access token
      description: Short-lived SET Service JWT access token. Refresh tokens are rejected by authenticated API middleware.
...
  responses:
...
    ServiceUnavailable:
      description: A required dependency, durable queue, scanner, or rate-limit store is unavailable.
      headers:
        X-Request-ID: { $ref: '#/components/headers/XRequestId' }
        X-Correlation-ID: { $ref: '#/components/headers/XCorrelationId' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Error:
      type: object
      required: [error, code, request_id, timestamp]
      properties:
        error: { type: string, example: Worker is not approved for login. }
        code: { type: string, example: WORKER_NOT_APPROVED }
        details: { type: object, nullable: true, example: { status: pending_approval } }
        request_id: { $ref: '#/components/schemas/RequestId' }
        timestamp: { type: string, format: date-time }
...
    WorkerRegisterRequest:
      type: object
      required: [full_name, phone, position]
      properties:
        full_name: { type: string, example: Ali Hasanov }
        phone: { type: string, example: "+994501234567" }
        position: { type: string, example: Waiter }
        skills:
          type: array
          items: { $ref: '#/components/schemas/Skill' }
        languages:
          type: array
          items: { type: string }
          example: [az, en]
    CompanyRegisterRequest:
      type: object
      required: [name, contact_name, email, phone]
      properties:
        name: { type: string, example: Grand Hotel Baku }
        contact_name: { type: string, example: Nigar Aliyeva }
        email: { type: string, format: email, example: company@setservice.az }
        phone: { type: string, example: "+994501234568" }
        docs_url: { type: string, format: uri }
        documents:
          type: array
          items: { $ref: '#/components/schemas/CompanyDocument' }
    RequestOtpRequest:
      type: object
      required: [phone]
      properties:
        phone: { type: string, example: "+994501234567" }
        purpose:
          type: string
          enum: [worker_registration, worker_password_reset]
          example: worker_registration
...
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token: { type: string }
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Compatibility registration endpoint for worker or company
      description: Prefer /auth/worker/register and /auth/company/register for new clients.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - allOf:
                    - $ref: '#/components/schemas/WorkerRegisterRequest'
                    - type: object
                      required: [role]
                      properties:
                        role: { type: string, enum: [worker] }
                - allOf:
                    - $ref: '#/components/schemas/CompanyRegisterRequest'
                    - type: object
                      required: [role]
                      properties:
                        role: { type: string, enum: [company] }
      responses:
        '201': { description: Account created and OTP sent }
  /auth/worker/register:
    post:
      tags: [Auth]
      summary: Register worker and send registration OTP
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkerRegisterRequest' }
      responses:
        '201': { description: Worker created with pending_otp status }
        '409':
          description: Phone already registered
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  /auth/worker/request-otp:
    post:
      tags: [Auth]
      summary: Request worker OTP for registration resend or password reset
      description: |
        Returns a generic anti-enumeration acknowledgement after the OTP has been accepted into the
        durable delivery flow. A successful HTTP response does not guarantee SMS delivery.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RequestOtpRequest' }
      responses:
        '200':
          description: OTP delivery request accepted or a generic anti-enumeration response returned
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OtpRequestAcceptedResponse' }
        '403':
          description: Worker is not approved or registration is not pending OTP
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /auth/worker/login:
    post:
      tags: [Auth]
      summary: Approved worker login with phone and password
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkerLoginRequest' }
      responses:
        '200':
          description: JWT token pair
          content: { application/json: { schema: { $ref: '#/components/schemas/TokenResponse' } } }
        '403':
          description: Worker is pending_approval, rejected, suspended, inactive, or unknown
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  /auth/worker/complete-registration:
    post:
      tags: [Auth]
      summary: Verify worker registration OTP and create password
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CompleteWorkerRegistrationRequest' }
      responses:
        '200': { description: Password set and worker moved to pending_approval }
        '401': { description: Invalid or expired OTP }
  /auth/worker/forgot-password:
    post:
      tags: [Auth]
      summary: Request worker password reset OTP
      description: |
        Always returns an anti-enumeration response. For an eligible account, delivery is queued
        asynchronously through the configured SMS or email provider.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkerForgotPasswordRequest' }
      responses:
        '200':
          description: Generic OTP delivery acknowledgement
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OtpRequestAcceptedResponse' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /auth/worker/reset-password:
    post:
      tags: [Auth]
      summary: Verify worker password reset OTP and set a new password
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkerResetPasswordRequest' }
      responses:
        '200': { description: Password reset }
        '401': { description: Invalid or expired OTP }
  /auth/worker/phone-change/request:
    post:
      tags: [Auth]
      security: [{ BearerAuth: [] }]
      summary: Request OTP for worker phone number change
      description: |
        The authenticated worker phone identity is not changed until OTP confirmation succeeds.
        Success means the OTP delivery was queued; it does not prove handset delivery.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkerPhoneChangeRequest' }
      responses:
        '200': { description: OTP sent to the new phone number }
        '409': { description: Phone number is already registered }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /auth/worker/phone-change/confirm:
    post:
      tags: [Auth]
      security: [{ BearerAuth: [] }]
      summary: Confirm worker phone number change with OTP
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkerPhoneChangeConfirmRequest' }
      responses:
        '200': { description: Worker phone identity updated }
        '401': { description: Invalid or expired OTP }
  /auth/email-verification/request:
    post:
      tags: [Auth]
      security: [{ BearerAuth: [] }]
      summary: Request email verification code for the authenticated user
      description: |
        The new email is stored as pending_email until code confirmation succeeds.
        Success means delivery was queued; it does not prove mailbox delivery.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EmailVerificationRequest' }
      responses:
        '200': { description: Email verification code sent or email already verified }
        '409': { description: Email is already registered }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /auth/email-verification/confirm:
    post:
      tags: [Auth]
      security: [{ BearerAuth: [] }]
      summary: Confirm pending email with verification code
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EmailVerificationConfirmRequest' }
      responses:
        '200': { description: Pending email promoted to verified email }
        '401': { description: Invalid or expired email verification code }
  /auth/company/register:
    post:
      tags: [Auth]
      summary: Register company and send OTP
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CompanyRegisterRequest' }
      responses:
        '201': { description: Company created with pending_approval status }
  /auth/company/login:
    post:
      tags: [Auth]
      summary: Approved company login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/EmailPasswordLoginRequest'
              example:
                email: company@setservice.az
                password: <explicit-company-password>
      responses:
        '200':
          description: JWT token pair
          content: { application/json: { schema: { $ref: '#/components/schemas/TokenResponse' } } }
        '403': { description: Company is not approved }
  /auth/company/web-login:
    post:
      tags: [Auth]
      security: []
      summary: Approved company browser login
      description: |
        Requires an exact trusted Origin. Returns the short-lived access token in the response
        and sets the rotating company refresh token in a role-scoped HttpOnly cookie.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EmailPasswordLoginRequest' }
      responses:
        '200':
          description: Browser session started
          headers:
            Set-Cookie:
              description: Role-scoped HttpOnly refresh cookie; Secure is mandatory for SameSite=None.
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WebSessionResponse' }
        '403': { description: Origin is not trusted or company is not approved }
  /auth/company/web-refresh:
    post:
      tags: [Auth]
      security: []
      summary: Rotate the company browser refresh cookie
      description: |
        Requires an exact trusted Origin and the company HttpOnly refresh cookie.
        The previous refresh token is revoked during rotation and is never returned in JSON.
      responses:
        '200':
          description: Browser session rotated
          headers:
            Set-Cookie:
              description: Newly rotated role-scoped HttpOnly refresh cookie.
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WebSessionResponse' }
        '401': { description: Browser session is missing, expired, or revoked }
        '403': { description: Origin is not trusted }
  /auth/company/web-logout:
    post:
      tags: [Auth]
      security: []
      summary: Revoke and clear the company browser session
      description: Requires an exact trusted Origin and revokes the refresh token from the HttpOnly cookie.
      responses:
        '204':
          description: Browser session revoked
          headers:
            Set-Cookie:
              description: Clears the role-scoped refresh cookie.
              schema: { type: string }
        '401': { description: Browser session is missing or expired }
        '403': { description: Origin is not trusted }
  /auth/company/complete-registration:
    post:
      tags: [Auth]
      summary: Verify company registration OTP and create password
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CompleteCompanyRegistrationRequest' }
      responses:
        '200': { description: Password set; company remains pending admin approval }
        '401': { description: Invalid or expired OTP }
  /auth/company/forgot-password:
    post:
      tags: [Auth]
      summary: Request company password reset OTP by email
      description: |
        Always returns an anti-enumeration response. For an eligible account, delivery is queued
        asynchronously through the configured email or SMS provider.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CompanyForgotPasswordRequest' }
      responses:
        '200':
          description: Generic OTP delivery acknowledgement
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OtpRequestAcceptedResponse' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /auth/company/reset-password:
    post:
      tags: [Auth]
      summary: Verify company password reset OTP and set a new password
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CompanyResetPasswordRequest' }
      responses:
        '200': { description: Password reset }
        '401': { description: Invalid or expired OTP }
  /auth/admin/login:
    post:
      tags: [Auth]
      summary: Super admin login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EmailPasswordLoginRequest' }
      responses:
        '200':
          description: JWT token pair
          content: { application/json: { schema: { $ref: '#/components/schemas/TokenResponse' } } }
  /auth/admin/web-login:
    post:
      tags: [Auth]
      security: []
      summary: Admin browser login
      description: |
        Requires an exact trusted Origin. Returns the short-lived access token in the response
        and sets the rotating admin refresh token in a role-scoped HttpOnly cookie.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EmailPasswordLoginRequest' }
      responses:
        '200':
          description: Browser session started
          headers:
            Set-Cookie:
              description: Role-scoped HttpOnly refresh cookie; Secure is mandatory for SameSite=None.
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WebSessionResponse' }
        '401': { description: Invalid credentials }
        '403': { description: Origin is not trusted }
  /auth/admin/web-refresh:
    post:
      tags: [Auth]
      security: []
      summary: Rotate the admin browser refresh cookie
      description: |
        Requires an exact trusted Origin and the admin HttpOnly refresh cookie.
        The previous refresh token is revoked during rotation and is never returned in JSON.
      responses:
        '200':
          description: Browser session rotated
          headers:
            Set-Cookie:
              description: Newly rotated role-scoped HttpOnly refresh cookie.
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WebSessionResponse' }
        '401': { description: Browser session is missing, expired, or revoked }
        '403': { description: Origin is not trusted }
  /auth/admin/web-logout:
    post:
      tags: [Auth]
      security: []
      summary: Revoke and clear the admin browser session
      description: Requires an exact trusted Origin and revokes the refresh token from the HttpOnly cookie.
      responses:
        '204':
          description: Browser session revoked
          headers:
            Set-Cookie:
              description: Clears the role-scoped refresh cookie.
              schema: { type: string }
        '401': { description: Browser session is missing or expired }
        '403': { description: Origin is not trusted }
  /auth/admin/forgot-password:
    post:
      tags: [Auth]
      deprecated: true
      summary: Disabled public admin password-reset endpoint
      description: |
        This public recovery flow is intentionally disabled and always returns
        `410 ADMIN_PASSWORD_RESET_DISABLED` without disclosing whether an admin account exists.
        Admin recovery must follow the separately authorized operational recovery runbook.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminForgotPasswordRequest' }
      responses:
        '410':
          description: Public admin password recovery is disabled (`ADMIN_PASSWORD_RESET_DISABLED`)
          headers:
            X-Request-ID: { $ref: '#/components/headers/XRequestId' }
            X-Correlation-ID: { $ref: '#/components/headers/XCorrelationId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example:
                error: Public admin password recovery is disabled.
                code: ADMIN_PASSWORD_RESET_DISABLED
                request_id: 7fa22d18-6e54-4c33-8154-9f68f8e2f84b
                timestamp: "2030-01-15T10:00:00.000Z"
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /auth/verify-otp:
    post:
      tags: [Auth]
      summary: Verify OTP for legacy registration/reset flows
      description: Password-based login is canonical. Login OTP purposes are deprecated and return OTP_LOGIN_DEPRECATED.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VerifyOtpRequest' }
      responses:
        '200':
          description: OTP verified
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/TokenResponse'
                  - type: object
                    properties:
                      user_id: { type: string, format: uuid }
                      worker_id: { type: string, format: uuid }
                      status: { $ref: '#/components/schemas/WorkerStatus' }
                      password_required: { type: boolean }
                      otp_challenge: { type: string }
        '401':
          description: Invalid or expired OTP
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '429':
          description: OTP blocked after failed attempts
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate refresh token and issue a new token pair
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshRequest' }
      responses:
        '200': { description: New token pair }
        '401': { description: Invalid refresh token }
  /auth/logout:
    post:
      tags: [Auth]
      security: [{ BearerAuth: [] }]
      summary: Revoke refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LogoutRequest' }
      responses:
        '204': { description: Logged out }
  /auth/fcm-token:
    post:
      tags: [Auth]
      security: [{ BearerAuth: [] }]
      summary: Register or update this device FCM token
      description: |
        Protected endpoint. The token is linked to the authenticated user and supports multiple devices per account.
        Use this after Worker, Company, or Admin login/refresh when Firebase returns a registration token.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FcmTokenRequest' }
            examples:
              android:
                value:
                  fcm_token: "fcm_registration_token_from_device"
                  platform: android
                  device_id: "android-installation-id"
      responses:
        '201':
          description: FCM token registered
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FcmTokenResponse' }
        '400': { description: Invalid token payload, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '401': { description: Bearer token is missing or invalid, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    patch:
      tags: [Auth]
      security: [{ BearerAuth: [] }]
      deprecated: true
      summary: Deprecated compatibility wrapper for registering this device FCM token
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FcmTokenRequest' }
      responses:
        '200':
          description: FCM token registered
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FcmTokenResponse' }
    delete:
      tags: [Auth]
      security: [{ BearerAuth: [] }]
      summary: Deactivate this device FCM token
      description: Call before logout when possible. Local logout should still clear tokens even if this request fails.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fcm_token]
              properties:
                fcm_token:
                  type: string
                  minLength: 10
                  maxLength: 4096
                  example: "fcm_registration_token_from_device"
      responses:
        '204': { description: FCM token deactivated }
        '400': { description: Invalid token payload, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '401': { description: Bearer token is missing or invalid, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...